Move run sanitizer tests in core folder
The original run_sanitizers.sh is moved to the `core` folder. By doing
so, we don't need to specify the `package` run in `cargo test` by `-p`,
and the paths to manifests `Cargo.toml` and `Cargo_ori.toml` is updated
as well.

The run_sanitizers.sh in the top folder now is just a proxy to call
`run_sanitizers.sh` in core folder.